Try requesting gzipped tiles.
This should have a couple of benefits:

1. Lower CPU usage in tilepacks, as we don't need to decompress the response.
2. Smaller mbtiles using compressed tiles.

However, there's a small issue; it seems that if CloudFront has a non-compressed tile in cache then it will return the tile without compressing it rather than fetch the compressed version from the origin. This means that we get back some uncompressed bytes and might need to compress them, but we don't know what the compression level is to match the settings on the origin, which might mean we end up not de-duplicating the tile when we could.